Update crates-io.md to not link to stale file
I was looking through the list and noticed several crates that I thought should not be there. After "git grep"ing the code I became convinced they should not have been there. Only when I wanted to make a PR to remove these stale entries, did I notice that the file in my editor had a very different list then the on linked here.

So to spare others, let's link to the master version of this.